Refactor ai-quirks to use a render hook system
