+++ pkg/backend/httpstate/client/ai_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package display
+package client
 
 import (
 	"encoding/json"
